Introduced decimals to TPCH queries and schema.
Former-commit-id: d0ee4cf0b4ecdf8c05da7e8b9e557b7b1d5475f4

+++ tpc-h/Queries/TPCH/Q1.hs
@@ -13,12 +13,7 @@ module Queries.TPCH.Q1
     ( q1
     ) where
 
-import qualified Prelude as P
 import Database.DSH
-import Database.DSH.Compiler
-
-import Database.HDBC.PostgreSQL
-
 import Schema.TPCH
 
 withFlagStatus :: Q LineItem -> Q (Text, Text)
@@ -30,17 +25,17 @@ filteredItems = [ li | li <- lineitems , l_shipdateQ li <= 42 ]
 fst9 :: (QA a, QA b, QA c, QA d, QA e, QA f, QA g, QA h, QA i) => Q (a, b, c, d, e, f, g, h, i) -> Q a
 fst9 (view -> (a, _, _, _, _, _, _, _, _)) = a
 
-q1 :: Q [((Text, Text), Double, Double, Double, Double, Double, Double, Double, Integer)]
-q1 = sortWith fst9 $ 
+q1 :: Q [((Text, Text), Decimal, Decimal, Decimal, Decimal, Decimal, Decimal, Decimal, Integer)]
+q1 = sortWith fst9 $
      [ tup9
-	  k
-	  (sum $ map l_quantityQ lis)
-	  (sum $ map l_extendedpriceQ lis)
-	  (sum $ map (\li -> l_extendedpriceQ li * (1 - l_discountQ li)) lis)
+          k
+          (sum $ map l_quantityQ lis)
+          (sum $ map l_extendedpriceQ lis)
+          (sum $ map (\li -> l_extendedpriceQ li * (1 - l_discountQ li)) lis)
 	  (sum $ map (\li -> l_extendedpriceQ li * (1 - l_discountQ li) * (1 + l_taxQ li)) lis)
-	  (avg $ map l_quantityQ lis)
-	  (avg $ map l_extendedpriceQ lis)
-	  (avg $ map l_discountQ lis)
-	  (length lis)
+          (avg $ map l_quantityQ lis)
+          (avg $ map l_extendedpriceQ lis)
+          (avg $ map l_discountQ lis)
+          (length lis)
       | (view -> (k, lis)) <- groupWithKey withFlagStatus filteredItems
       ]
