+++ tpc-h/Queries/TPCH/Q18.hs
@@ -8,34 +8,29 @@
 {-# LANGUAGE TypeFamilies          #-}
 {-# LANGUAGE UndecidableInstances  #-}
 {-# LANGUAGE ViewPatterns          #-}
-    
+
 -- TPC-H Q18
 
 module Queries.TPCH.Q18
     ( q18
     ) where
 
-import qualified Prelude as P
 import Database.DSH
-import Database.DSH.Compiler
-
-import Database.HDBC.PostgreSQL
-
 import Schema.TPCH
 
-largeOrders :: Double -> Q [Integer]
-largeOrders quantity = 
+largeOrders :: Decimal -> Q [Integer]
+largeOrders quantity =
   [ ok
   | (view -> (ok, ls)) <- groupWithKey l_orderkeyQ lineitems
   , toQ quantity < sum (map l_quantityQ ls)
   ]
 
-sortSpec :: Q ((Text, Integer, Integer, Integer, Double), Double) -> Q (Double, Integer)
+sortSpec :: Q ((Text, Integer, Integer, Integer, Decimal), Decimal) -> Q (Decimal, Integer)
 sortSpec gs =
   let (view -> (_, _, _, orderDate, totalPrice)) = fst gs
   in pair (-1 * totalPrice) orderDate
 
-q18 :: Double -> Q [((Text, Integer, Integer, Integer, Double), Double)]
+q18 :: Decimal -> Q [((Text, Integer, Integer, Integer, Decimal), Decimal)]
 q18 quantity =
   sortWith sortSpec $
   map (\(view -> (k, g)) -> pair k (sum $ map snd g)) $
