+++ tpc-h/Queries/TPCH/Q22.hs
@@ -8,32 +8,25 @@
 {-# LANGUAGE TypeFamilies          #-}
 {-# LANGUAGE UndecidableInstances  #-}
 {-# LANGUAGE ViewPatterns          #-}
-    
+
 -- TPC-H Q21
 
 module Queries.TPCH.Q22
     ( q22
     ) where
 
-import qualified Data.Text as T
-
-import qualified Prelude as P
 import Database.DSH
-import Database.DSH.Compiler
-
-import Database.HDBC.PostgreSQL
-
 import Schema.TPCH
 
-customersAvgBalance :: [Text] -> Q Double
+customersAvgBalance :: [Text] -> Q Decimal
 customersAvgBalance areaPrefixes =
-  avg [ c_acctbalQ c 
+  avg [ c_acctbalQ c
       | c <- customers
       , c_acctbalQ c > 0
-      , subString 1 2 (c_phoneQ c) `elem` toQ areaPrefixes 
+      , subString 1 2 (c_phoneQ c) `elem` toQ areaPrefixes
       ]
 
-potentialCustomers :: [Text] -> Q [(Text, Double)]
+potentialCustomers :: [Text] -> Q [(Text, Decimal)]
 potentialCustomers areaPrefixes =
   [ pair (subString 1 2 (c_phoneQ c)) (c_acctbalQ c)
   | c <- customers
@@ -42,7 +35,7 @@ potentialCustomers areaPrefixes =
   , null [ 1 :: Q Integer | o <- orders, o_custkeyQ o == c_custkeyQ c ]
   ]
 
-potentialCustomers' :: [Text] -> Q [(Text, Double)]
+potentialCustomers' :: [Text] -> Q [(Text, Decimal)]
 potentialCustomers' areaPrefixes =
   [ pair (subString 1 2 (c_phoneQ c)) (c_acctbalQ c)
   | c <- customers
@@ -51,11 +44,11 @@ potentialCustomers' areaPrefixes =
   , c_custkeyQ c `notElem` (map o_custkeyQ orders)
   ]
 
-q22 :: [Text] -> Q [(Text, Integer, Double)]
-q22 areaPrefixes = 
+q22 :: [Text] -> Q [(Text, Integer, Decimal)]
+q22 areaPrefixes =
   sortWith (\(view -> (c, _, _)) -> c) $
   [ tup3 cntrycode
            (length pas)
-	   (sum $ map snd pas)
+           (sum $ map snd pas)
   | (view -> (cntrycode, pas)) <- groupWithKey fst (potentialCustomers areaPrefixes)
   ]
